CMT #2.1: Started implementation of progress tracker.
Basic idea is working. Some tweaking here and there is needed to make it fully functional.

Current idea is:
Keep the track of current test's average and at the end of the test, store it in a file. Also fetch the average of all the average scores of test prior to current test. Compare the current test with previous averages and show if there's an improvement or not. Currently the only calculation is of the time taken to solve each question.
Late this will store a far more condensed report.