CMT #4: Generating truly random questions now
The program keeps a track of operands generated in each section by storing it in a set. Whenever a operand is generated, its existence is then checked within the previous generated set. If found, the operand is generated again otherwise program runs its normal course.

std::set is used to keep track of previously generated questions and is chosen just because of its avg O(n) time complexity for find operation.