CMT #8.51: Smart Bots Implementation
I know this isn't actually a bot because it doesn't have any actual 'smartness' in it. It contains a few predefined operations that 'mimic' smart behavior. Since I do not know what should I call them instead of 'bot' I'm just calling them bot. I'm open to suggestions and currently considering calling it 'Chichi'. Whatever.

Chichi follows a simple logic, bah! it isn't even a logic to be honest. That is defined in a commented section of code present in get_bot_input( ). It has a delibrate slowdown system and degree of inaccuracy variable that are both controlled by a global variable.

And just to be cool, I've changed the record format from lame .dat to .greym
Might change it later.