RESTRUCT_06: Added private methods skeleton.
Created empty files for containing private methods implementation of classes Walnut and Test respectively. Furthermore, created a basic constructor file for Chara and in the process of doing it I realized that Test folder should be inside the Walnut folder because Test class is defined inside the Walnut class. Will look into it later.