MAIN 9 : Improved Timer implementation syntax
Defined a Stopwatch class and created a variable of its type in the Test class that will be used throughout the project. It is basically a wrapper around previous existing timing procedure.
It separated the Timing section into a separate entity that can be fixed easily and quickly. Improved program syntax and debugging for future commits. Briefly tested.